+++ services/rfq/relayer/relconfig/iconfig_generated.go
@@ -3,6 +3,9 @@
 package relconfig
 
 import (
+	"math/big"
+
+	"github.com/ethereum/go-ethereum/common"
 	"github.com/synapsecns/sanguine/ethergo/signer/config"
 )
 
@@ -40,4 +43,9 @@ type IConfig interface {
 	GetDestGasEstimate(chainID uint32) int
 	// GetL1FeeParams returns the L1 fee params for the given chain.
 	GetL1FeeParams(chainID uint32, origin bool) (uint32, int, bool)
+	// GetQuotePct returns the quote percentage.
+	GetQuotePct() float64
+	// GetMinQuoteAmount returns the quote amount for the given chain and address.
+	// Note that this getter returns the value in native token decimals.
+	GetMinQuoteAmount(chainID int, addr common.Address) *big.Int
 }