+++ packages/synapse-interface/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.237](https://github.com/synapsecns/sanguine/compare/@synapsecns/synapse-interface@0.1.236...@synapsecns/synapse-interface@0.1.237) (2024-01-12)
+
+**Note:** Version bump only for package @synapsecns/synapse-interface
+
+
+
+
+
+## [0.1.236](https://github.com/synapsecns/sanguine/compare/@synapsecns/synapse-interface@0.1.235...@synapsecns/synapse-interface@0.1.236) (2024-01-12)
+
+**Note:** Version bump only for package @synapsecns/synapse-interface
+
+
+
+
+
 ## [0.1.235](https://github.com/synapsecns/sanguine/compare/@synapsecns/synapse-interface@0.1.234...@synapsecns/synapse-interface@0.1.235) (2024-01-11)
 
 **Note:** Version bump only for package @synapsecns/synapse-interface